chore: rename CW CP contracts with cosmwasm prefix
