+++ packages/testcontainers/src/cosmwasm-artifacts.ts
