+++ packages/testcontainers/src/index.ts
@@ -1,3 +1,3 @@
 export { AnvilContainer, ChainName, saltToHex, StartedAnvilContainer } from "./anvil-container";
-export { CosmWasmContainer, StartedCosmWasmContainer } from "./container";
-export { Contract, SatLayerContracts } from "./satlayer";
+export { CosmWasmContract, CosmWasmContracts } from "./cosmwasm";
+export { CosmWasmContainer, StartedCosmWasmContainer } from "./wasmd-container";
